STADTPULS-327: todo: Could we make postgrest accept our tokens?
This would allow us to provide a full rest api based on postgrest.
If a user wants programmatic access he currently needs to have

- the anon token as header -> apikey
- a user with password to login

with that he can get a user token and make requests to our supabaseUrl.
He will have to refesh his token every hour.
For less friction we could have a route that:

- adds the apikey
- passes everything else through to supabaseUrl/rest/v1

Then we have the full crud api withhout wrapping it and doing the grunt work