fix(pagination): Fix nextPage field in reply body
The URL was build using the request.url. If the URL had query params
the fields offset and limit just got appended and then where duplicated
this made all requests fail that loop through the reponses took the
nextPage url as the indicator if there are more records.
The result was that it only worked to make two iterations. Since all the
tests only checked one iteration this was not seen.