CORS, pagination and syslog - staging to main (#69, #70, #75, #76) (#77)
* feature/STADTPULS 415/add pino syslog (#69)

* STADTPULS-415: housekeeping

* STADTPULS-415: feat(logging): Enable logging in syslog style

* fix(CORS): Set  Access-Control-Allow-Origin (#75)

The header was not explicitly set and it made some requests fail

* fix(pagination): Fix nextPage field in reply body

The URL was build using the request.url. If the URL had query params
the fields offset and limit just got appended and then where duplicated
this made all requests fail that loop through the reponses took the
nextPage url as the indicator if there are more records.
The result was that it only worked to make two iterations. Since all the
tests only checked one iteration this was not seen.

* fix(pagination): Fix nextPage field in reply body (#76)

* fix(syslog newline): Fixes newline@ syslog end (#70)

Since the options are not typed I added an s to newline