feat(crypto): Adds own hash and compare functions
This adds a new crypto module with a hashing and compare function.
These are used to create new hashes for identifying tokens in our db.
The bycrypt module we used before did only create hashes based on
the first 72bytes of a string. This is more then enough for passwords
but our tokens where about 300bytes long. Since the beginning of the
string was the same for many tokens we had false positive matches
when comparing incomning tokens with the stored hashes.
Now we should have unique matches