fix/STADTPULS 470/verify multi auth token flow (#88)
* feat(crypto): Adds own hash and compare functions

This adds a new crypto module with a hashing and compare function.
These are used to create new hashes for identifying tokens in our db.
The bycrypt module we used before did only create hashes based on
the first 72bytes of a string. This is more then enough for passwords
but our tokens where about 300bytes long. Since the beginning of the
string was the same for many tokens we had false positive matches
when comparing incomning tokens with the stored hashes.
Now we should have unique matches

* housekeeping

* feat(crypto): Add token check to ttn

* some requests for development

* housekeeping

* store inspect exec for nodemon

* incr max rows in example .envs

* formatting and todo

* STADTPULS-470: test(crypto): Adds unit and integration tests

fixes a typo

* STADTPULS-470: test(MAX ROWS): exports max rows from env

* STADTPULS-470: fix(test): Random word seems to make tests fail

* STADTPULS-470: logging

* STADTPULS-470: housekeeping

* STADTPULS-470: playing with payloads from ttn. WIP

* STADTPULS-470: test(snapshot): Update snapshots

* STADTPULS-470: fix(snapshot)

* STADTPULS-470: test(token): Posting records with multiple tokens

* STADTPULS-470: housekeeping

* STADTPULS-470: fix(test): URL for ttn needs to be integration...

My own design shot me in the foot